Fix 'Config Your App' example config
The yaml example and json example is said to be the same, but defines components with different names. Yaml file defines `main` while JSON file defines `frontend. 

In the break down, there is again a mixture of the two component names. It states that `frontend` is defined, but in the example url for the component, it again referres to `main`.  According to what we learned in chapter [Radix Concepts](https://www.radix.equinor.com/start/radix-concepts/#component) the url is built up from the component name, and since `frontend` seems to be the 'winning' component name in the example, the url should reflect that.